LayoutTests/imported/w3c/web-platform-tests/dom/events/Event-propagation-expected.txt


PASS Newly-created Event 
FAIL After stopPropagation() assert_equals: Propagation flag after first dispatch expected true but got false
PASS Reinitialized after stopPropagation() 
FAIL After stopImmediatePropagation() assert_equals: Propagation flag after first dispatch expected true but got false
PASS Reinitialized after stopImmediatePropagation() 
